+++ CHANGELOG.md
@@ -2,10 +2,30 @@
 
 All notable changes of RelicFrog Raspy PI Firewall release series are documented in this file using the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-_This RelicFrog/Jenkins project changelog documentation start with version 1.0.0 (2020-05-14)_
+_This RelicFrog/Gatekeeper project changelog documentation start with version 1.0.0 (2020-05-14)_
+
+
+## [1.0.1], 2020-05-15:
+_current_
+
+### Added
+
+* root exec check
+* error handling method
+* wlan/eth ip identifier
+* core tool stack checker
+* minor changes in documentation
+
+### Removed
+
+* show method (obsolete)
+
+### Fixed
+
+* error in input chain handler (missing ip)
+
 
 ## [1.0.0], 2020-05-14:
-_current_
 
 ### Added